Move split-related Tree History describes into history_tree_split_test
Pull every split- and merge-related describe out of
history_tree_test.ts and place them in history_tree_split_test.ts
so the file's name matches its content. Sections moved:

- Single Client - Split/Merge (was 4)
- Single Client - Split L1 undo/redo (was 4b)
- Single Client - Split L1 chained ops (was 4c)
- Multi Client - Split L1 convergence (was 4d)
- Split L1 edge cases (was 4e)
- Single Client - Split L2 undo/redo (was 4f)
- Single Client - Split L2 chained ops (was 4g)
- Multi Client - Split L2 convergence (was 4h)
- Multi Client - Split L2 edge cases (was 4i)

The pre-tombstoned descendant filter describe stays in this file
as section 5. The remaining basic / chained / edge / multi-client
basic / reconcile / multi-client edge / style sections remain in
history_tree_test.ts; the comment numbering for the style sections
is updated from 5/6/7 to 4/5/6 to keep numbering contiguous.

Both files share the same `xmlOf` helper (one-line). The
multi-client split sections import `withTwoClientsAndDocuments`
from the integration helper. Test counts are preserved exactly:
133 + 78 = 211 passing tests, 5 + 1 skipped.
